Arreglado el resetPassword y getFavorites con los eventos filtrados
